requirements.txt
jsonschema==3.2.0
PyYAML==5.3.1
python-gitlab==2.6.0
